Implement debug hasel support (#287)
* Implement hasel/hawindow support.

This should allow simultaneous resume and halt to work.

* Fix anyrunning/anyhalted bits.

* Add --without-hasel argument for testing.

* Make halt/resume times more equal.

Switching threads after every instruction executed in debug mode leads
to a lot of extra instructions being executed on the "other" thread when
both are really supposed to halt/resume near-simultaneously. Fixed that
by adding wfi to debug_rom.S, and implementing it to switch to the other
hart as well as check for JTAG input.

When resuming, write the hart ID to the debug ROM so that the DM knows
which hart actually resumed. (Before simultaneous resume it just assumed
the current one.)

Also got rid of resume symbol in debug_rom.S since it had no purpose.

* Preserve Debug ROM entry points.

* Make sure minstret is correct when wfi happens.