Extends the commit log feature with memory writes. (#324)
* Extends the commit log feature with memory writes.

This provides a little more information for debugging instruction
traces, allowing you to maintain the state of memory as the trace
is processed.

The following sample trace output illustrates the formatting of
the new memory writes. The first line is an instruction at
location 0x80000094, containing the bytes (0x80830313) and
commiting the value 0x80000898 to the register x6. The second
line is an instruction which neither commits a register nor
writes memory. The third line writes the value 0x0 to
0x80000890.

    3 0x80000094 (0x80830313) x 6 0x80000898
    3 0x80000098 (0x0062d663)
    3 0x8000009c (0x00028023) mem 0x80000890 0x0

* Changes addressing feedback from review.